requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy<2.0.0
Pillow>=9.0.0
tqdm>=4.65.0
scikit-learn>=0.24.2
matplotlib>=3.4.3
pandas>=1.3.3
pillow>=8.3.2
pytest>=6.2.5
black>=21.9b0
flake8>=3.9.2 